app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.25.0
pytest>=7,<9
requests>=2.31.0
